% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{address_to_gps}
\alias{address_to_gps}
\title{address_to_gps : Convertir une adresse en coordonnées GPS}
\usage{
address_to_gps(address)
}
\arguments{
\item{address}{Adresse en caractères.}
}
\value{
Coordonnées GPS.
}
\description{
address_to_gps : Convertir une adresse en coordonnées GPS
}
