% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.character}
\alias{get_forecast.character}
\title{get_forecast.character : Obtenir des prévisions météorologiques basées sur une adresse en caractères}
\usage{
get_forecast.character(address)
}
\arguments{
\item{address}{Adresse en caractères.}
}
\value{
Un tibble avec des données de prévisions météorologiques.
}
\description{
get_forecast.character : Obtenir des prévisions météorologiques basées sur une adresse en caractères
}
