% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast.numeric}
\alias{get_forecast.numeric}
\title{get_forecast.numeric: Obtenir des prévisions météorologiques basées sur des coordonnées numériques}
\usage{
get_forecast.numeric(coordinate)
}
\arguments{
\item{coordinate}{Vecteur numérique de coordonnées.}
}
\value{
Un tibble avec des données de prévisions météorologiques.
}
\description{
get_forecast.numeric: Obtenir des prévisions météorologiques basées sur des coordonnées numériques
}
