% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_forecast}
\alias{get_forecast}
\title{get_forecast : Obtenir des prévisions météorologiques en fonction d'une adresse ou de coordonnées numériques}
\usage{
get_forecast(x)
}
\arguments{
\item{x}{Adresse en caractères ou vecteur numérique de coordonnées.}
}
\value{
Un tibble avec des données de prévisions météorologiques.
}
\description{
get_forecast : Obtenir des prévisions météorologiques en fonction d'une adresse ou de coordonnées numériques
}
\examples{
get_forecast(c(47.83706, -4.34687))
get_forecast("Paris, France")

}
