% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{get_gps_coordinate}
\alias{get_gps_coordinate}
\title{get_gps_coordinate: Obtenir les coordonnées GPS à partir d'une adresse}
\usage{
get_gps_coordinate(address)
}
\arguments{
\item{address}{Adresse en caractères.}
}
\value{
Coordonnées GPS.
}
\description{
get_gps_coordinate: Obtenir les coordonnées GPS à partir d'une adresse
}
