% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_forecast.R
\name{unnest_response}
\alias{unnest_response}
\title{unnest_response}
\usage{
unnest_response(perf7)
}
\arguments{
\item{perf7}{Données de prévisions météorologiques.}
}
\value{
Un tibble avec les données de prévisions météorologiques.
}
\description{
Extrait les données de température en fonction de l'heure.
}
